Split party model class file
